feat: improve output for list and upload commands
This adds respectively a long output option (-l/--long) to show the date
and number of chunks of files, and a progress output showing how many
chunks are needed and how many are already uploaded.
Another addition is a new upload command argument (--star-chunk) allowing
to start uploading part of a file from a specific chunk index instead of
the start, in case an upload failed or was cancelled.